requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
google-genai==0.1.0
pydantic==2.4.2
python-multipart==0.0.6